Breaking change: extension and renaming in data_ingest
This only breaks if InputOutputData was used directly in user code, it then has
to be renamed.

Rationale: For tensor models we will often deal with arrays and it is useful
to split them with the current data splitter. This usage come in the next commits.

To reach that goal this commit does the following

1) Introduces a base class for IOData, the splitter can be applied to all implementations
2) Adds a new class InputOutputDataArrays
3) Renamed InputOutputData to InputOutputDataFrames <-- This is the breaking change